% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dmisc2-package.R
\docType{package}
\name{Dmisc2-package}
\alias{Dmisc2-package}
\title{Dieter's misc functions
 
 For latex-related functions of the old Dmisc package, see Dlatex}
\description{
\tabular{ll}{ Package: \tab Dmisc\cr Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2006-01-03\cr License: \tab MIT\cr }
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}
\keyword{misc}
\keyword{print}
