% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMlatticeOptions.R
\name{DMlatticeOptions}
\alias{DMlatticeOptions}
\title{Dieter's default lattice options}
\usage{
DMlatticeOptions(superpose.polygon.col = NULL,
  superpose.line.col = NULL, par.strip.text.cex = NULL,
  transparent.strip = FALSE)
}
\arguments{
\item{superpose.polygon.col}{a palette for polygon colors, e.g.
\code{brewer.pal(5,"Set1")}}

\item{superpose.line.col}{a palette for line colors, e.g.
\code{brewer.pal(8,"Spectral")}}

\item{par.strip.text.cex}{text size of strips}

\item{transparent.strip}{If true, make panel strips transparent}
}
\description{
Set default lattice options, e.g. for use by Sweave.
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}
