% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorPalettes.R
\name{DMPalette}
\alias{DMPalette}
\title{Special palettes}
\usage{
DMPalette(palette = NULL)
}
\arguments{
\item{palette}{Name of palette. When omitted, returns names of available palettess}
}
\description{
Some fixed palettes
}

