% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContrasts.r
\name{getContrasts}
\alias{getContrasts}
\alias{readContrasts}
\title{Read and format contrast tables from Excel}
\usage{
getContrasts(cname, excelfile, rows = NULL)

readContrasts(cname, excelfile)
}
\arguments{
\item{cname}{Sheet name in Excel file}

\item{excelfile}{path of Excel file with contrast table}

\item{rows}{optionally only select some rows of the table}
}
\value{
\code{readContrasts} read a contrast table. Only columns with at
least on period  in the header are kept.
\code{getContrasts} additionally includes column and row name
information to label contrast tables with estimable (gmodels)
}
\description{
Reads contrast from an Excel file for use with \code{estimable}
of package gmodels.
}
\examples{
library(nlme)
library(gmodels)
options(digits=3)
set.seed(4711)
excelfile = system.file("extdata", "contrasts.xlsx", package = "Dmisc2")
d = expand.grid(subject = LETTERS[1:8],
                peri= c("Wine", "Tea"),
                post = c("Water", "Kirsch"),
                interval = c("Pre", "Post"))
d$vol = round(rnorm(nrow(d),10,2),1)
d.lme = lme(vol~interval+peri+post+peri:interval+interval:post,
            data=d,random=~1|subject)
summary(d.lme)
ct = getContrasts("peripostinterval",excelfile)
estimable(d.lme,ct,conf.int=0.95)
}
\seealso{
See \code{constrast.xlsx} in the extdata directory for some
rules of thumb on constructing a contrast matrix.
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}
\keyword{models}
