% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDIofMCMC.R
\name{HDIofMCMC}
\alias{HDIofMCMC}
\title{Highest density interval HDI from a sample}
\usage{
HDIofMCMC(sampleVec, credMass = 0.95)
}
\arguments{
\item{sampleVec}{a vector of representative values from a probability distribution.}

\item{credMass}{a scalar between 0 and 1, indicating the mass within the credible
interval that is to be estimated.}
}
\value{
a vector containing the limits of the HDI
}
\description{
Computes highest density interval from a sample of representative values,
  estimated as shortest credible interval.
}
\examples{
set.seed(1)
HDIofMCMC(rnorm(100))
}
\author{
John K. Kruschke
}

