% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packet_panel_bycolumn.R
\name{packet.panel.bycolumn}
\alias{packet.panel.bycolumn}
\title{lattice panel layout for multipage consistency}
\usage{
packet.panel.bycolumn(layout, condlevels, page, row, column, skip)
}
\arguments{
\item{layout}{the layout argument in high level functions, suitably standardized.}

\item{condlevels}{a list of levels of conditioning variables, after relevant permutations and/or reordering of levels}

\item{page}{location of panel on page}

\item{row}{location of panel in row}

\item{column}{the location of the panel in the coordinate system columns}

\item{skip}{the skip argument in high level functions}
}
\description{
Gabor Grothendiek's solution in 
\url{http://stackoverflow.com/questions/9654244/multipage-lattice-panel-arrangement}
}
\examples{
library(lattice)
library(latticeExtra)
d <- expand.grid(f1 = as.factor(letters[1:8]),
                 f2 = as.factor(LETTERS[1:3]),
                 x  = 0:10)
d$y <- rnorm(nrow(d))
p <- xyplot(y~x|f1+f2, data=d, cex=0.5, pch=16)
p <- useOuterStrips(p)
p <- update(p, layout=c(5,3)) # plotting from here on gives a warning
suppressWarnings(plot(p, packet.panel=packet.panel.bycolumn))
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}
\seealso{
\code{\link[lattice]{packet.panel.default}}
}

