% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairsdf.R
\name{pairsdf}
\alias{pairsdf}
\title{Pairs plot for mixed data sets}
\usage{
pairsdf(df, abbr = TRUE, abbr.len = 4)
}
\arguments{
\item{df}{dataframe with numericals and factor data}

\item{abbr}{allow abbreviated labels}

\item{abbr.len}{lenght of abbreviated labels}
}
\description{
A variant of the pairs plot to create a 
scatterplot matrix, allowing for categorical variables.
}
\examples{
require(MASS)
par(mar=c(0,0,0,0)) # Much less whitespace
par(mar=c(1,1,1,1)) # Much less whitespace
pairsdf(coop[,1:3])
pairsdf(farms[,1:3])
par(mar=c(1,1,1,1))
pairsdf(coop[,2:4])
}
\author{
Biostatmat, \url{http://biostatmatt.com/archives/2398}
}
