% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorPalettes.R
\name{pal}
\alias{pal}
\title{Plotting palettes}
\usage{
pal(palette, name = "")
}
\arguments{
\item{palette}{a palette}

\item{name}{x axis label}
}
\description{
List of palettes
}
\examples{
pal(DMPalette("greenmono"))
Palettes = DMPalette()
par(mfrow=c(4,4))
lapply(Palettes,function(x) pal(DMPalette(x)))
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}
