% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveNA.r
\name{RemoveNAColumns}
\alias{RemoveNAColumns}
\alias{RemoveNARows}
\title{Removes NA columns or rows in a data frame}
\usage{
RemoveNAColumns(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
A data frame with removed columns or rows that are all-NA removed
}
\description{
RemoveNARows is different from na.omit, because it removes only rows that
are completely NA
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}
\keyword{misc}

