% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_dir_create.R
\name{safe.dir.create}
\alias{safe.dir.create}
\title{Creates a directory if it does not exist}
\usage{
safe.dir.create(path)
}
\arguments{
\item{path}{path to the directory to be created}
}
\value{
TRUE if the directory was created, FALSE if it already existed,
exception if the directory could not be created
}
\description{
Raises if creation was not successful
}

